# Optimize complexity of client endpoint WeightedRoundRobinStrategy