--- a/core/src/main/java/com/linecorp/armeria/client/endpoint/WeightedRoundRobinStrategy.java
+++ b/core/src/main/java/com/linecorp/armeria/client/endpoint/WeightedRoundRobinStrategy.java
@@ -16,8 +16,8 @@
 
 package com.linecorp.armeria.client.endpoint;
 
+import java.util.ArrayList;
 import java.util.Comparator;
-import java.util.LinkedList;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
@@ -71,82 +71,123 @@ final class WeightedRoundRobinStrategy implements EndpointSelectionStrategy {
             return endpointsAndWeights.selectEndpoint(currentSequence);
         }
 
+        // endpoints accumulation which are grouped by weight
+        private static final class EndpointsGroupByWeight {
+            final long startIndex;
+            final int weight;
+            final long accumulatedWeight;
+
+            EndpointsGroupByWeight(long startIndex, int weight, long accumulatedWeight) {
+                this.startIndex = startIndex;
+                this.weight = weight;
+                this.accumulatedWeight = accumulatedWeight;
+            }
+        }
+
+        //
+        // In general, assume the weights are w0 < w1 < ... < wM where M = N - 1, N is number of endpoints.
+        //
+        // * The first part of result: (a0..aM)(a0..aM)...(a0..aM) [w0 times for N elements].
+        // * The second part of result: (a1..aM)...(a1..aM) [w1 - w0 times for N - 1 elements].
+        // * and so on
+        //
+        // In this way:
+        //
+        // * Total number of elements of first part is: X(0) = w0 * N.
+        // * Total number of elements of second part is: X(1) = (w1 - w0) * (N - 1)
+        // * and so on
+        //
+        // Therefore, to find endpoint for a sequence S = currentSequence % totalWeight, firstly we find
+        // the part which sequence belongs, and then modular by the number of elements in this part.
+        //
+        // Accumulation function F:
+        //
+        // * F(0) = X(0)
+        // * F(1) = X(0) + X(1)
+        // * F(2) = X(0) + X(1) + X(2)
+        // * F(i) = F(i-1) + X(i)
+        //
+        // We could easily find the part (which sequence S belongs) using binary search on F.
+        // Just find the index k where:
+        //
+        //                               F(k) <= S < F(k + 1).
+        //
+        // So, S belongs to part number (k + 1), index of the sequence in this part is P = S - F(k).
+        // Because part (k + 1) start at index (k + 1), and contains (N - k - 1) elements,
+        // then the real index is:
+        //
+        //                              (k + 1) + (P % (N - k - 1))
+        //
+        // For special case like w(i) == w(i+1). We just group them all together
+        // and mark the start index of the group.
+        //
         private static final class EndpointsAndWeights {
             private final List<Endpoint> endpoints;
             private final boolean weighted;
             private final long totalWeight; // prevent overflow by using long
-
-            private static final class EndpointsGroupByWeight {
-                int startIndex;
-                int weight;
-                long accumulatedWeight;
-
-                EndpointsGroupByWeight(int startIndex, int weight, long accumulatedWeight) {
-                    this.startIndex = startIndex;
-                    this.weight = weight;
-                    this.accumulatedWeight = accumulatedWeight;
-                }
-            }
-
             private final EndpointsGroupByWeight[] endpointsGroupByWeight;
 
             EndpointsAndWeights(Iterable<Endpoint> endpoints) {
+                long totalWeight = 0;
                 int minWeight = Integer.MAX_VALUE;
                 int maxWeight = Integer.MIN_VALUE;
-                long totalWeight = 0;
+                int numberDistinctWeight = 0;
 
-                // get min and max weight
-                for (Endpoint endpoint : endpoints) {
-                    final int weight = endpoint.weight();
-                    minWeight = Math.min(minWeight, weight);
-                    maxWeight = Math.max(maxWeight, weight);
-                }
-
-                // prepare endpoints
-                List<Endpoint> endps = ImmutableList.copyOf(endpoints)
+                // prepare endpoints by copying
+                this.endpoints = ImmutableList.copyOf(endpoints)
                         .stream()
-                        .filter(endpoint -> endpoint.weight() > 0) // only process endpoint with weight > 0
+                        .filter(e -> e.weight() > 0) // only process endpoint with weight > 0
                         .sorted(Comparator
                                 .comparing(Endpoint::weight)
                                 .thenComparing(Endpoint::host)
                                 .thenComparingInt(Endpoint::port))
-                        .collect(Collectors.toList());
-                int numEndpoints = endps.size();
+                        .collect(Collectors.toCollection(ArrayList::new));
+                long numEndpoints = this.endpoints.size();
+
+                // get min weight, max weight and number of distinct weight
+                int oldWeight = -1;
+                for (Endpoint endpoint : this.endpoints) {
+                    final int weight = endpoint.weight();
+                    minWeight = Math.min(minWeight, weight);
+                    maxWeight = Math.max(maxWeight, weight);
+                    numberDistinctWeight += weight == oldWeight ? 0 : 1;
+                    oldWeight = weight;
+                }
 
                 // accumulation
-                LinkedList<EndpointsGroupByWeight> accumulatedGroups = new LinkedList<>();
+                EndpointsGroupByWeight[] accumulatedGroups = new EndpointsGroupByWeight[numberDistinctWeight];
                 EndpointsGroupByWeight currentGroup = null;
-                int rest = numEndpoints;
-                for (Endpoint endpoint : endps) {
+                int index = -1;
+
+                long rest = numEndpoints;
+                for (Endpoint endpoint : this.endpoints) {
                     if (currentGroup == null || currentGroup.weight != endpoint.weight()) {
                         totalWeight += currentGroup == null ?
-                                (long) endpoint.weight() * (long) rest
-                                : (long) (endpoint.weight() - currentGroup.weight) * (long) rest;
-                        currentGroup = new EndpointsGroupByWeight(numEndpoints - rest,
-                                endpoint.weight(), totalWeight);
-                        accumulatedGroups.addLast(currentGroup);
+                                (long) endpoint.weight() * rest
+                                : (long) (endpoint.weight() - currentGroup.weight) * rest;
+                        currentGroup = new EndpointsGroupByWeight(
+                                numEndpoints - rest, endpoint.weight(), totalWeight
+                        );
+                        accumulatedGroups[++index] = currentGroup;
                     }
 
                     rest--;
                 }
 
-                this.endpoints = endps;
-                this.endpointsGroupByWeight = accumulatedGroups.toArray(
-                        new EndpointsGroupByWeight[accumulatedGroups.size()]
-                );
+                this.endpointsGroupByWeight = accumulatedGroups;
                 this.totalWeight = totalWeight;
                 this.weighted = minWeight != maxWeight;
             }
 
-            Endpoint selectEndpoint(int currentSequence) {
+            Endpoint selectEndpoint(long currentSequence) {
                 if (endpoints.isEmpty()) {
                     throw new EndpointGroupException(endpoints + " is empty");
                 }
 
-                int numberEndpoints = endpoints.size();
+                long numberEndpoints = endpoints.size();
 
                 if (weighted) {
-                    long mod = Math.abs((long) (currentSequence) % totalWeight);
+                    long mod = Math.abs(currentSequence % totalWeight);
 
                     if (mod < endpointsGroupByWeight[0].accumulatedWeight) {
                         return endpoints.get((int) (mod % numberEndpoints));
@@ -170,13 +211,13 @@ final class WeightedRoundRobinStrategy implements EndpointSelectionStrategy {
                     }
 
                     // (left + 1) is the part where sequence belongs
-                    int indexInPart = (int) (mod - endpointsGroupByWeight[left].accumulatedWeight);
-                    int realIndex = endpointsGroupByWeight[left + 1].startIndex +
+                    long indexInPart = mod - endpointsGroupByWeight[left].accumulatedWeight;
+                    long realIndex = endpointsGroupByWeight[left + 1].startIndex +
                             indexInPart % (numberEndpoints - endpointsGroupByWeight[left + 1].startIndex);
-                    return endpoints.get(realIndex);
+                    return endpoints.get((int) realIndex);
                 }
 
-                return endpoints.get(Math.abs(currentSequence % numberEndpoints));
+                return endpoints.get((int) Math.abs(currentSequence % numberEndpoints));
             }
         }
     }