Motivation:

Client using EndpointGroup with current WeightedRoundRobinStrategy is slow. 
* Current algorithm complexity for selectEndpoint operation is O(N * MaxWeight) where N is number of endpoints. Arbitrary weight value (>1000 for example) cause it slow. 

Modifications:
- Sort endpoints by weight, then by host and port.
- Find the correct endpoint faster by binary search on accumulated sum of weight.
- The complexity is O(log(N)) and not depend on weight value.

Algorithm details:
- In comment code block

Result:
* Faster selectEndpoint operation.

Benchmarks:

Before
```
Result "com.linecorp.armeria.core.client.endpoint.WeightedRoundRobinStrategyBenchmark.randomMainly1Max30":
  392327.218 ±(99.9%) 9226.013 ops/s [Average]

Result "com.linecorp.armeria.core.client.endpoint.WeightedRoundRobinStrategyBenchmark.randomMax10":
  289799.381 ±(99.9%) 3761.818 ops/s [Average]

Result "com.linecorp.armeria.core.client.endpoint.WeightedRoundRobinStrategyBenchmark.randomMax100":
  75949.628 ±(99.9%) 1970.120 ops/s [Average]

Result "com.linecorp.armeria.core.client.endpoint.WeightedRoundRobinStrategyBenchmark.sameWeight":
  54412957.573 ±(99.9%) 1253690.454 ops/s [Average]

Result "com.linecorp.armeria.core.client.endpoint.WeightedRoundRobinStrategyBenchmark.unique":
  21152.272 ±(99.9%) 11486.626 ops/s [Average]
```

After
```
Result "com.linecorp.armeria.core.client.endpoint.WeightedRoundRobinStrategyBenchmark.randomMainly1Max30":
  28586820.346 ±(99.9%) 145391.823 ops/s [Average]

Result "com.linecorp.armeria.core.client.endpoint.WeightedRoundRobinStrategyBenchmark.randomMax10":
  25691640.444 ±(99.9%) 574921.960 ops/s [Average]

Result "com.linecorp.armeria.core.client.endpoint.WeightedRoundRobinStrategyBenchmark.randomMax100":
  24246013.498 ±(99.9%) 994566.750 ops/s [Average]

Result "com.linecorp.armeria.core.client.endpoint.WeightedRoundRobinStrategyBenchmark.sameWeight":
  54149030.441 ±(99.9%) 1580181.334 ops/s [Average]

Result "com.linecorp.armeria.core.client.endpoint.WeightedRoundRobinStrategyBenchmark.unique":
  21813513.256 ±(99.9%) 1958592.548 ops/s [Average]
```