@linxGnu 2018-06-12
Thank you very much @imasahiro @anuraaga @trustin 
I have updated source code and added JMH as your suggestion. Benchmark is also included in comment.
Please take a review.